requirements.txt
pytest-playwright
pytest-html
pytest-xdist
python-dotenv
requests
